refactor spring.boot to springboot package (#1497)
