+++ shenyu-spring-boot-starter/shenyu-spring-boot-starter-plugin/shenyu-spring-boot-starter-plugin-alibaba-dubbo/src/main/resources/META-INF/spring.factories
@@ -16,4 +16,4 @@
 #
 
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.apache.shenyu.spring.boot.starter.plugin.alibaba.dubbo.AlibabaDubboPluginConfiguration
+org.apache.shenyu.springboot.starter.plugin.alibaba.dubbo.AlibabaDubboPluginConfiguration
