+++ shenyu-spring-boot-starter/shenyu-spring-boot-starter-sync-data-center/shenyu-spring-boot-starter-sync-data-websocket/src/main/resources/META-INF/spring.factories
@@ -16,4 +16,4 @@
 #
 
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.apache.shenyu.spring.boot.starter.sync.data.websocket.WebsocketSyncDataConfiguration
+org.apache.shenyu.springboot.starter.sync.data.websocket.WebsocketSyncDataConfiguration
