--- shenyu-spring-boot-starter/shenyu-spring-boot-starter-sync-data-center/shenyu-spring-boot-starter-sync-data-zookeeper/src/main/java/org/apache/shenyu/spring/boot/sync/data/zookeeper/ZookeeperSyncDataConfiguration.java
@@ -1,74 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.shenyu.spring.boot.sync.data.zookeeper;
-
-import lombok.extern.slf4j.Slf4j;
-import org.I0Itec.zkclient.ZkClient;
-import org.apache.shenyu.sync.data.api.AuthDataSubscriber;
-import org.apache.shenyu.sync.data.api.MetaDataSubscriber;
-import org.apache.shenyu.sync.data.api.PluginDataSubscriber;
-import org.apache.shenyu.sync.data.api.SyncDataService;
-import org.apache.shenyu.sync.data.zookeeper.ZookeeperSyncDataService;
-import org.springframework.beans.factory.ObjectProvider;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import java.util.Collections;
-import java.util.List;
-
-/**
- * Zookeeper sync data configuration for spring boot.
- */
-@Configuration
-@ConditionalOnClass(ZookeeperSyncDataService.class)
-@ConditionalOnProperty(prefix = "shenyu.sync.zookeeper", name = "url")
-@EnableConfigurationProperties(ZookeeperConfig.class)
-@Slf4j
-public class ZookeeperSyncDataConfiguration {
-
-    /**
-     * Sync data service sync data service.
-     *
-     * @param zkClient          the zk client
-     * @param pluginSubscriber the plugin subscriber
-     * @param metaSubscribers   the meta subscribers
-     * @param authSubscribers   the auth subscribers
-     * @return the sync data service
-     */
-    @Bean
-    public SyncDataService syncDataService(final ObjectProvider<ZkClient> zkClient, final ObjectProvider<PluginDataSubscriber> pluginSubscriber,
-                                           final ObjectProvider<List<MetaDataSubscriber>> metaSubscribers, final ObjectProvider<List<AuthDataSubscriber>> authSubscribers) {
-        log.info("you use zookeeper sync shenyu data.......");
-        return new ZookeeperSyncDataService(zkClient.getIfAvailable(), pluginSubscriber.getIfAvailable(),
-                metaSubscribers.getIfAvailable(Collections::emptyList), authSubscribers.getIfAvailable(Collections::emptyList));
-    }
-
-    /**
-     * register zkClient in spring ioc.
-     *
-     * @param zookeeperConfig the zookeeper configuration
-     * @return ZkClient {@linkplain ZkClient}
-     */
-    @Bean
-    public ZkClient zkClient(final ZookeeperConfig zookeeperConfig) {
-        return new ZkClient(zookeeperConfig.getUrl(), zookeeperConfig.getSessionTimeout(), zookeeperConfig.getConnectionTimeout());
-    }
-}
