+++ shenyu-spring-boot-starter/shenyu-spring-boot-starter-sync-data-center/shenyu-spring-boot-starter-sync-data-zookeeper/src/main/resources/META-INF/spring.factories
@@ -16,4 +16,4 @@
 #
 
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.apache.shenyu.spring.boot.sync.data.zookeeper.ZookeeperSyncDataConfiguration
+org.apache.shenyu.springboot.sync.data.zookeeper.ZookeeperSyncDataConfiguration
